refactor: move tryFunctions to core and include tests
