+++ packages/core/src/amazonq/lsp/util.ts
@@ -26,20 +26,3 @@ export async function lspSetupStage<T>(
         return result
     })
 }
-
-/**
- * Try Functions in the order presented and return the first returned result. If none return, throw the final error.
- * @param functions non-empty list of functions to try.
- * @returns
- */
-export async function tryFunctions<Result>(functions: (() => Promise<Result>)[]): Promise<Result> {
-    let currentError: Error = new Error('No functions provided')
-    for (const func of functions) {
-        try {
-            return await func()
-        } catch (e) {
-            currentError = e as Error
-        }
-    }
-    throw currentError
-}
